nvelles fonctions supplementaires

+++ libft/ft_striteri.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void		ft_striteri(char *s, void (*f)(size_t, char *))
+void		ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i])
