+++ libft/ft_strmapi.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-char		*ft_strmapi(char const *s, char (*f)(size_t, char))
+char		*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	char	*result;
 	int		i;
